+++ client/src/action/createSubmitAction.ts
@@ -1,4 +1,4 @@
-import action from '..';
+import action from '.';
 
 
 interface CreateSubmitConfig {
